refactor(components): fix imports
